refactor start route
