+++ src/handlers/menu/handler.rs
@@ -7,7 +7,7 @@ use teloxide::{
 
 use crate::{
     handlers::state::MenuCommandState,
-    routes::{command_help, command_invite, command_start},
+    routes::{command_help, command_invite, start::command_start},
 };
 
 use super::MenuCommandsHandler;
@@ -25,7 +25,7 @@ impl<'a: 'static> MenuCommandsHandler<'a>
             Update::filter_message().branch(
                 dptree::entry()
                     .filter_command::<MenuCommandState>()
-                    .branch(dptree::case![MenuCommandState::Start].endpoint(command_start::command))
+                    .branch(dptree::case![MenuCommandState::Start].endpoint(command_start))
                     .branch(dptree::case![MenuCommandState::Help].endpoint(command_help::command))
                     .branch(
                         dptree::case![MenuCommandState::Invite].endpoint(command_invite::command),
