+++ src/handlers/menu/mod.rs
@@ -1,5 +1,4 @@
 mod handler;
-mod tests;
 
 pub trait MenuCommandsHandler<'a> {
     fn setup_main_commands(self) -> Self;
