--- src/handlers/menu/model.rs
@@ -1,1 +1,0 @@
-
