+++ src/routes/mod.rs
@@ -2,5 +2,5 @@ pub mod callback_invite_query;
 pub mod command_help;
 pub mod command_invite;
 pub mod command_open_main_menu;
-pub mod command_start;
 pub mod model;
+pub mod start;
